getPos() added
Spider & Butterfly: returns a float array[]
index 0 = x, 1=y, 2=z;

NOTE: centipede returns a two dimensional array. This is because each
body segment has it's own coordinates.
so [0][0], would be the head's X coordinate. This is in addition to the
new method centipede.bodyLength() which will return the number of
visible body segments.  Thus the "tail" of the centipede can be accessed
like:
//returns the Z coordinate of the last segment of the first centipede.
centi[0].getPos()[body.length()-1][2];